+++ src/main/java/domain/LottoTickets.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.strategy.GenerateStrategy;
 import java.util.Collections;
 import java.util.List;
 import java.util.Set;
