+++ src/main/java/domain/dto/WinningResultDto.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.dto;
 
 public class WinningResultDto {
     private final int matchCount;
