+++ src/main/java/domain/strategy/GenerateStrategy.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.strategy;
 
 import java.util.Set;
