+++ src/main/java/domain/strategy/WinningPrizeStrategy.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.strategy;
+
+import domain.WinningPrize;
 
 public interface WinningPrizeStrategy {
     WinningPrize findWinningPrize(int matchCount, boolean matchBonus);
