+++ src/main/java/view/OutputView.java
@@ -2,8 +2,8 @@ package view;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import domain.LottoTicketDto;
-import domain.WinningResultDto;
+import domain.dto.LottoTicketDto;
+import domain.dto.WinningResultDto;
 
 public class OutputView {
     private static final String PURCHASE_COUNT_MESSAGE = "수동으로 %d장, 자동으로 %d개를 구매했습니다.\n";
