+++ src/test/java/domain/LottoGameTest.java
@@ -2,6 +2,9 @@ package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.strategy.GenerateStrategy;
+import domain.strategy.LottoWinningPrizeStrategy;
+import domain.strategy.WinningPrizeStrategy;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashSet;
