+++ src/test/java/domain/LottoTicketsTest.java
@@ -3,6 +3,7 @@ package domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import domain.strategy.GenerateStrategy;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashSet;
